requirements.txt
numpy
pytesseract
pygetwindow
pyautogui
opencv-python
pynput
python_Levenshtein
rapidfuzz
playsound